Ändra typen för initialZebraCount och initialCheetahCount från byte till int för att kunan ha fler djur Fixa utskrifterna ännu lite till

+++ src/com/jcalm/BoardFactory.java
@@ -14,7 +14,7 @@ public class BoardFactory {
         board = new Board();
     } // privat konstruktor
 
-    public static void createBoard(int size, byte initialZebraCount, byte initialCheetahCount) {
+    public static void createBoard(int size, int initialZebraCount, int initialCheetahCount) {
         board = new Board(size, initialZebraCount, initialCheetahCount);
     }
